server/requirements.txt
flask
python-dateutil
flask_prometheus_metrics
elasticsearch
